+++ youtui/src/app/keycommand.rs
@@ -80,7 +80,7 @@ impl Keybind {
 
 impl<A: Action> Display for KeyCommand<A> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let w: String = 
+        let w: String =
             // NOTE: Replace with standard library method once stabilised.
             itertools::intersperse(
                 self