+++ json-crawler/CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.0.3](https://github.com/nick42d/youtui/compare/json-crawler/v0.0.2...json-crawler/v0.0.3) - 2024-10-26
+
+### Fixed
+- [**breaking**] Make album optional for songs search. Closes [#174](https://github.com/nick42d/youtui/pull/174) ([#176](https://github.com/nick42d/youtui/pull/176))
+- _album field on SearchResultSong is now optional, removed public ways to create custom error from json_crawler (CrawlerError::array_size_from_context, JsonCrawlerIterator::get_context and JsonCrawlerArrayIterContext struct)_ 
+
 ## [0.0.2](https://github.com/nick42d/youtui/compare/json-crawler/v0.0.1...json-crawler/v0.0.2) - 2024-09-04
 
 ### Other
 - Update dependencies ([#155](https://github.com/nick42d/youtui/pull/155))
-
-
-
 
 ## [0.0.1](https://github.com/nick42d/youtui/releases/tag/json-crawler/v0.0.1) - 2024-08-12